Move some functionality from caches into new base_caches module, remove clear_caches from backend_interface
