+++ python/tlm_adjoint/equations.py
@@ -32,10 +32,11 @@ from .backend_code_generator_interface import assemble, \
     process_solver_parameters, r0_space, rhs_addto, rhs_copy, solve, \
     update_parameters_dict, verify_assembly
 
+from .base_caches import CacheRef
 from .base_equations import AssignmentSolver, Equation, EquationException, \
     NullSolver, get_tangent_linear, no_replace_compatibility
-from .caches import CacheRef, assembly_cache, form_neg, is_cached, \
-    linear_solver_cache, split_form
+from .caches import assembly_cache, form_neg, is_cached, linear_solver_cache, \
+    split_form
 from .functions import bcs_is_cached, bcs_is_homogeneous, bcs_is_static, \
     eliminate_zeros, extract_coefficients, is_r0_function
