+++ python/tlm_adjoint/hessian.py
@@ -21,8 +21,8 @@
 from .interface import function_copy, function_get_values, \
     function_is_cached, function_is_checkpointed, function_is_static, \
     function_name
-from .backend_interface import clear_caches
 
+from .base_caches import clear_caches
 from .manager import manager as _manager, set_manager
 
 __all__ = \
